+++ src/main/java/org/teamseven/tetris/handler/GameHandler.java
@@ -1,0 +1,12 @@
+package org.teamseven.tetris.handler;
+
+import org.teamseven.tetris.score.Score;
+
+public class GameHandler {
+    private boolean paused;
+    private Score score;
+
+    public boolean isPaused() {
+        return paused;
+    }
+}
